+++ tests/unit/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+@pytest.fixture()
+def mocker(request, mocker):
+    request.cls.mocker = mocker
+    return mocker